[AN] refactor: 검색 결과 화면 SearchViewModel 사용 #482
