+++ android/app/src/main/java/com/onair/hearit/presentation/search/SearchViewModelFactory.kt
@@ -3,12 +3,22 @@ package com.onair.hearit.presentation.search
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.onair.hearit.di.RepositoryProvider
+import com.onair.hearit.di.UseCaseProvider
+import com.onair.hearit.domain.model.SearchInput
 
 @Suppress("UNCHECKED_CAST")
-class SearchViewModelFactory : ViewModelProvider.Factory {
+class SearchViewModelFactory(
+    private val input: SearchInput?,
+) : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
         val categoryRepository = RepositoryProvider.categoryRepository
         val recentKeywordRepository = RepositoryProvider.recentKeywordRepository
-        return SearchViewModel(categoryRepository, recentKeywordRepository) as T
+        val getSearchResultUseCase = UseCaseProvider.getSearchResultUseCase
+        return SearchViewModel(
+            categoryRepository,
+            recentKeywordRepository,
+            getSearchResultUseCase,
+            input,
+        ) as T
     }
 }
